requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
boto3==1.10.3
botocore==1.13.3
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.4
django-admin-interface==0.10.5
django-allauth==0.40.0
django-colorfield==0.1.15
django-cors-headers==3.1.0
django-extensions==2.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-modeladmin-reorder==0.3.1
django-rest-auth==0.9.5
django-storages==1.7.2
djangorestframework==3.10.2
docutils==0.15.2
Faker==2.0.1
gunicorn==20.0.4
hvac==0.10.0
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
